Repro for vscode-docker debugging issue.
When trying to debug a .NET project with its `OutputPath` property set
to an absolute path, debugging into the docker container appears to fail
with an "application <path> does not exist" error.  This attempts to
create a minimal reproduction, where the only step required to switch
between working as expected and not is commenting out the `OutputPath`
property in the csproj.

After debugging into the vscode-docker extension, the path in the error
message matches the `containerAppOutput` value created in
`NetCoreDebugHelper.resolveLaunchDebugConfiguration` (in
vscode-docker:/src/debugging/netcore/NetCoreDebugHelper.ts). I am
unable to debug into the full VSCode app, so I can't follow that value
through to where it's actually sent to the debugger in the container,
so I can't actually verify that's the source of the issue.